requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
Pillow==9.2.0
psycopg2==2.9.5
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
